+++ app/Models/Chapter.php
@@ -8,4 +8,13 @@ use Illuminate\Database\Eloquent\Model;
 class Chapter extends Model
 {
     use HasFactory;
+
+    // Relationships.
+    public function book () {
+        return $this->belongsTo(Book::class);
+    }
+
+    public function verses () {
+        return $this->hasMany(Verse::class);
+    }
 }