+++ app/Models/Verse.php
@@ -6,5 +6,8 @@ use Illuminate\Database\Eloquent\Model;
 
 class Verse extends Model
 {
-    //
+    // Relationships.
+    public function chapter () {
+        return $this->belongsTo(Chapter::class);
+    }
 }