+++ routes/web.php
@@ -32,6 +32,9 @@ Route::get('/home', 'HomeController@index')->name('home');
 
 Route::get('/dashboard', 'Admin\DashboardController@index') -> name('dashboard') -> middleware('can:admin');
 
+Route::get('/search', 'Bible\SearchBibleController@index' );
+Route::get('/verse', 'Bible\SearchVerseController@index');
+
 Auth::routes();
 
 Route::get('/email/verify', function () {